Moves yaml data into the layouts module
